lazy load integration and nvml
