+++ kernel_tuner/util.py
@@ -41,8 +41,6 @@ def _get_cupy():
         return None
     return _cp
 
-from kernel_tuner.observers.nvml import NVMLObserver
-
 # number of special values to insert when a configuration cannot be measured
 
 
@@ -223,6 +221,7 @@ def check_tune_params_list(tune_params, observers, simulation_mode=False):
         if name in forbidden_names:
             raise ValueError("Tune parameter " + name + " with value " + str(param) + " has a forbidden name!")
     if any("nvml_" in param for param in tune_params):
+        from kernel_tuner.observers.nvml import NVMLObserver
         if not simulation_mode and (not observers or not any(isinstance(obs, NVMLObserver) for obs in observers)):
             raise ValueError("Tune parameters starting with nvml_ require an NVMLObserver!")
 
@@ -429,6 +428,36 @@ def get_best_config(results, objective, objective_higher_is_better=False):
         key=lambda x: x[objective] if isinstance(x[objective], float) else ignore_val,
     )
     return best_config
+
+
+# specifies for a number of pre-defined objectives whether
+# the objective should be minimized or maximized (boolean value denotes higher is better)
+objective_default_map = {
+    "time": False,
+    "energy": False,
+    "fitness": True,
+    "cost": False,
+    "loss": False,
+    "GFLOP/s": True,
+    "TFLOP/s": True,
+    "GB/s": True,
+    "TB/s": True,
+    "GFLOPS/W": True,
+    "TFLOPS/W": True,
+    "GFLOP/J": True,
+    "TFLOP/J": True,
+}
+
+
+def get_objective_defaults(objective, objective_higher_is_better):
+    """Use time as default objective and infer objective_higher_is_better for known objectives."""
+    objective = objective or "time"
+    if objective_higher_is_better is None:
+        if objective in objective_default_map:
+            objective_higher_is_better = objective_default_map[objective]
+        else:
+            raise ValueError(f"Please specify objective_higher_is_better for objective {objective}")
+    return objective, objective_higher_is_better
 
 
 def get_config_string(params, keys=None, units=None):
